requirements.txt
absl-py==0.9.0
asgiref==3.2.7
astunparse==1.6.3
backtrader==1.9.74.123
beautifulsoup4==4.8.2
bs4==0.0.1
cachetools==4.1.0
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
Django==3.0.2
gast==0.3.3
google-auth==1.13.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.28.1
h5py==2.10.0
idna==2.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lxml==4.4.2
Markdown==3.2.1
matplotlib==3.1.2
mplfinance==0.12.3a3
multitasking==0.0.9
numpy==1.17.4
oauthlib==3.1.0
opt-einsum==3.2.0
pandas==1.0.1
pandas-datareader==0.8.1
Pillow==7.0.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyMySQL==0.9.3
pymysql-pooling==1.0.6
pyparsing==2.4.5
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
selenium==3.141.0
six==1.13.0
slacker==0.13.0
soupsieve==1.9.5
sqlparse==0.3.0
tensorboard==2.2.0
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.2.0rc2
tensorflow-estimator==2.2.0rc0
termcolor==1.1.0
urllib3==1.25.7
Werkzeug==1.0.1
wrapt==1.12.1
yfinance==0.1.54